// src/shared/Models/Const/Permissions.enum.ts
export enum PERMISSIONS {
  AUTHORIZATION = 'AUTHORIZATION',
  GRANT_ACCESS = 'GRANT_ACCESS',
  EDIT_USER = 'EDIT_USER',
  GET_CONTRACTS_OWN = 'GET_CONTRACTS_OWN',
  GET_CONTRACTS_REGION = 'GET_CONTRACTS_REGION',
  GET_CONTRACTS_ALL = 'GET_CONTRACTS_ALL',
  VIEW_ARCHIVE = 'VIEW_ARCHIVE',
  WITHDRAWAL = 'WITHDRAWAL',
  TRANSFER_TO_PARKING = 'TRANSFER_TO_PARKING',
  REQUEST_FOR_EVALUATION = 'REQUEST_FOR_EVALUATION',
  INSPECTION_CHECK = 'INSPECTION_CHECK',
  RECEIPT_OF_CLAIM = 'RECEIPT_OF_CLAIM',
  GET_PARKINGS_OWN = 'GET_PARKINGS_OWN',
  GET_PARKINGS_REGION = 'GET_PARKINGS_REGION',
  GET_PARKINGS_ALL = 'GET_PARKINGS_ALL',
  CAR_CONTROL = 'CAR_CONTROL',
  CAR_CONTROL_OWN = 'CAR_CONTROL_OWN',
  CAR_CONTROL_ALL = 'CAR_CONTROL_ALL',
  EVALUATION_HANDLE = 'EVALUATION_HANDLE',
  AUTOMATIC_EVALUATION = 'AUTOMATIC_EVALUATION',
  STATISTIC_LEGAL_COUNSEL = 'STATISTIC_LEGAL_COUNSEL',
  STATISTIC_SUPERVISOR = 'STATISTIC_SUPERVISOR',
  STATISTIC_DIRECTOR = 'STATISTIC_DIRECTOR',
  STATISTIC_APPRAISER = 'STATISTIC_APPRAISER',
  STATISTIC_VALIDATOR = 'STATISTIC_VALIDATOR',
  GET_USER_LIST = 'GET_USER_LIST',
  BLOCK_USER = 'BLOCK_USER',
  EDIT_DICTIONARIES = 'EDIT_DICTIONARIES',
  VIEW_LOGS = 'VIEW_LOGS',
  RECEIPT_OF_NOTE = 'RECEIPT_OF_NOTE',
  STATISTICS_INSPECTION = 'STATISTICS_INSPECTION',
  STATISTICS_CONTROL = 'STATISTICS_CONTROL',
  STATISTICS_EVALUATION =  'STATISTICS_EVALUATION',
  MACRO_REGION_HEAD = 'MACRO_REGION_HEAD',
}